+++ swing/common-ui/src/main/java/is/codion/swing/common/ui/laf/LookAndFeelComboBox.java
@@ -90,7 +90,8 @@ public final class LookAndFeelComboBox extends JComboBox<Item<LookAndFeelEnabler
 		if (builder.enabled != null) {
 			enabled(builder.enabled, this);
 		}
-		includeInstalled = State.builder(builder.includeInstalled)
+		includeInstalled = State.builder()
+						.value(builder.includeInstalled)
 						.listener(getModel().items()::filter)
 						.build();
 		getModel().items().visible().predicate()